pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = [
    {name = "capprinha",email = "gutoklein21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bookstore.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

